fix for assertion error in `text_extract_cell_dataframe`
Error was caused by different datatypes ('int32' vs 'int64') for the Panda DataFrame: enforcing 'int64' for the assertion sample:

>       pdt.assert_frame_equal(
            segmentation.extract_cell_dataframe(cell_labels),
            expected_from_labels)
E       AssertionError: Attributes of DataFrame.iloc[:, 0] (column name="area") are different
E
E       Attribute "dtype" are different
E       [left]:  int64
E       [right]: int32

mibidata\tests\test_segmentation.py:193: AssertionError

+++ mibidata/tests/test_segmentation.py
@@ -189,7 +189,7 @@ class TestSegmentation(unittest.TestCase):
         expected_from_labels = pd.DataFrame(
             np.array([areas, x_centroids, y_centroids]).T,
             columns=['area', 'x_centroid', 'y_centroid'],
-            index=pd.Index(labels, name='label'))
+            index=pd.Index(labels, name='label'), dtype='int64')
         pdt.assert_frame_equal(
             segmentation.extract_cell_dataframe(cell_labels),
             expected_from_labels)
